overview: fix "internal error: no overview data found" on 32 bit architectures - fixes #9723
The precision field in the backend overview YAML files can hold
fs.ModTimeNotSupported (100 years in nanoseconds) which overflows int
on 32 bit platforms, making the YAML for those backends fail to parse
and causing rclone to log 18 internal errors on every invocation.

Use int64 for the precision field and add a test that parses every
embedded backend YAML file so this is caught on 32 bit test runs.

(cherry picked from commit 5629f2668c69149bf3d9d8e2a25bb32a2648606e)

+++ backend/overview/overview.go
@@ -26,7 +26,7 @@ type BackendConfig struct {
 	Remote           string   `yaml:"remote"`
 	Features         []string `yaml:"features"`
 	Hashes           []string `yaml:"hashes"`
-	Precision        int      `yaml:"precision"`
+	Precision        int64    `yaml:"precision"`
 }
 
 // GetBackendConfig from docs/data/backends
